+++ src/pages/[nickname]/page.tsx
@@ -1,101 +1,16 @@
-import { useEffect } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { MarkingThumbnailGrid } from "@/widgets/marking/ui";
-import { ProfileOverView, EmptyProfileOverView } from "@/widgets/profile/ui";
-import { TemporaryMarkingBar } from "@/entities/marking/ui";
-import { useGetMyFollowingIdsMap, useGetProfile } from "@/entities/profile/api";
-import { ROUTER_PATH } from "@/shared/constants";
 import { useNicknameParams } from "@/shared/lib/profile";
-import { useAuthStore } from "@/shared/store";
-import { SettingIcon } from "@/shared/ui/icon";
-import {
-  BackwardNavigationBar,
-  NavigationBar,
-} from "@/shared/ui/navigationbar";
-import { NotFoundUser } from "./notFoundUser";
+import { MyProfileDashboard } from "./myProfileDashboard";
+import { OtherProfileDashboard } from "./otherProfileDashboard";
 
 /**
  * 해당 컴포넌트는 /:nickname 경로로 들어온 사용자의 프로필 페이지를 나타냅니다.
  * @:nickname: 해당 닉네임을 가진 사용자의 프로필 페이지, 만약 AuthStore에 저장된 닉네임과 같다면 마이페이지처럼 이용 가능 합니다.
  */
 export const ProfilePage = () => {
-  const { nicknameParams, isMyPage } = useNicknameParams();
-  const token = useAuthStore((state) => state.token);
-  const navigate = useNavigate();
+  const { isMyPage } = useNicknameParams();
 
-  const { data, isError, error } = useGetProfile({
-    nickname: nicknameParams,
-  });
-  const { data: myFollowingIdsMap } = useGetMyFollowingIdsMap();
-
-  // TODO 404 와 같은 양식의 디자인 사용하는건 어떤지 디자이너와 상의
-  useEffect(() => {
-    if (!token) {
-      navigate(ROUTER_PATH.LOGIN);
-    }
-  }, [token, navigate]);
-
-  if (isError && error.code === 404) {
-    return <NotFoundUser />;
+  if (isMyPage) {
+    return <MyProfileDashboard />;
   }
-
-  if (!data || !myFollowingIdsMap) {
-    return;
-  }
-
-  const { followersIds, followingsIds, pet, tempCnt, userId } = data;
-
-  return (
-    <>
-      {isMyPage ? (
-        <MyPageNavigationBar />
-      ) : (
-        <BackwardNavigationBar
-          label={<h1 className="text-grey-900 title-1">{nicknameParams}님</h1>}
-        />
-      )}
-      <section className="px-4 flex flex-col items-start gap-8">
-        {pet ? (
-          <ProfileOverView
-            nickname={nicknameParams}
-            followersIds={followersIds}
-            followingsIds={followingsIds}
-            isFollowing={myFollowingIdsMap[userId]}
-            pet={pet}
-          />
-        ) : (
-          <EmptyProfileOverView isMyPage={isMyPage} />
-        )}
-        <div className="flex flex-col items-start gap-2 w-full ">
-          <h3 className="text-grey-900 text-center title-2">
-            {isMyPage ? "내 마킹" : `${nicknameParams}님의 마킹`}
-          </h3>
-          {isMyPage && typeof tempCnt === "number" && tempCnt > 0 && (
-            <TemporaryMarkingBar tempCnt={tempCnt} />
-          )}
-          {/* {nickname} 유저의 마킹들 3x3 썸네일 */}
-          <MarkingThumbnailGrid nickname={nicknameParams} />
-        </div>
-      </section>
-    </>
-  );
-};
-
-const MyPageNavigationBar = () => {
-  const nickname = useAuthStore((state) => state.nickname);
-  return (
-    <NavigationBar
-      componentType="buttonRight"
-      label={<h1 className="text-grey-900 title-1">{nickname}님</h1>}
-      button={
-        <Link
-          to={ROUTER_PATH.SETTING}
-          className="px-3 py-3 text-grey-500"
-          aria-label="내 정보 설정하기"
-        >
-          <SettingIcon />
-        </Link>
-      }
-    />
-  );
+  return <OtherProfileDashboard />;
 };
