+++ src/widgets/profile/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./profileOverview";
+export * from "./myProfileOverview";
